move convertWatchKindToContent to separate file
