+++ internal/services/v1/watch_test.go
@@ -16,8 +16,6 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	v1svc "github.com/authzed/spicedb/internal/services/v1"
-
 	"github.com/authzed/spicedb/internal/datastore/memdb"
 	"github.com/authzed/spicedb/internal/testfixtures"
 	"github.com/authzed/spicedb/internal/testserver"
@@ -432,57 +430,6 @@ definition document {
 				_, err := stream.Recv()
 				grpcutil.RequireStatus(t, tc.expectedCode, err)
 			}
-		})
-	}
-}
-
-func TestConvertWatchKindToContent(t *testing.T) {
-	tests := []struct {
-		name  string
-		kinds []v1.WatchKind
-		exp   datastore.WatchContent
-	}{
-		{
-			name:  "default",
-			kinds: []v1.WatchKind{},
-			exp:   datastore.WatchRelationships,
-		},
-		{
-			name:  "unspecified",
-			kinds: []v1.WatchKind{v1.WatchKind_WATCH_KIND_UNSPECIFIED},
-			exp:   datastore.WatchRelationships,
-		},
-		{
-			name:  "only_relationship_updates",
-			kinds: []v1.WatchKind{v1.WatchKind_WATCH_KIND_INCLUDE_RELATIONSHIP_UPDATES},
-			exp:   datastore.WatchRelationships,
-		},
-		{
-			name:  "relationships_and_schema_updates",
-			kinds: []v1.WatchKind{v1.WatchKind_WATCH_KIND_INCLUDE_SCHEMA_UPDATES},
-			exp:   datastore.WatchRelationships | datastore.WatchSchema,
-		},
-		{
-			name:  "relationships_and_checkpoints",
-			kinds: []v1.WatchKind{v1.WatchKind_WATCH_KIND_INCLUDE_CHECKPOINTS},
-			exp:   datastore.WatchRelationships | datastore.WatchCheckpoints,
-		},
-		{
-			name: "all",
-			kinds: []v1.WatchKind{
-				v1.WatchKind_WATCH_KIND_INCLUDE_RELATIONSHIP_UPDATES,
-				v1.WatchKind_WATCH_KIND_INCLUDE_SCHEMA_UPDATES,
-				v1.WatchKind_WATCH_KIND_INCLUDE_CHECKPOINTS,
-			},
-			exp: datastore.WatchRelationships | datastore.WatchSchema | datastore.WatchCheckpoints,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			require := require.New(t)
-			result := v1svc.ConvertWatchKindToContent(tt.kinds)
-			require.Equal(tt.exp, result)
 		})
 	}
 }
